Improve DiscreteLocator implementation
This refactors internal implementation of DiscreteLocator
so that it is applied for any BoundaryNorm or DiscreteNorm
passed to a colorbar func rather than depending on a hidden
guide_kw 'locator' value. This also helps simplify use of
level centers for tick values over level edges. And finally
it moves _parse_discrete_norm from Axes to PlotAxes.